Update SpecialAttack system with new features
- Add random_target field to special_attacks table
- Change value column type in special_attack_effects to decimal for float support
- Add new effect types: HEALING_PER_TURN, ADD_ARMOR, ADD_RESISTANCE_PERCENT
- Add FROST element to SpecialAttackElement enum
- Update SpecialAttack and SpecialAttackEffect models
- Update SpecialAttackSeeder with new attacks and NPCs (Łowczyni Wspomnień, Przyzwacz Demonów, Maddok Magua, Tezcatlipoca, Tanroth)
- Create new migration files for backward compatibility

+++ app/Models/SpecialAttackEffect.php
@@ -20,7 +20,7 @@ class SpecialAttackEffect extends DynamicModel
 
     protected $casts = [
         'type' => SpecialAttackEffectType::class,
-        'value' => 'integer',
+        'value' => 'float',
         'duration' => 'integer',
     ];
